Working on doc and wrapper for  Maxent.

+++ Analysis/Max_SAC_ph.f90
@@ -1,35 +1,58 @@
      Program Test
 
        Use MaxEnt_stoch_mod
-  
-       Implicit Real (Kind=Kind(0.d0)) (A-G,O-Z)
-       Implicit Integer (H-N)
+
+       Implicit  None
+       !Implicit Real (Kind=Kind(0.d0)) (A-G,O-Z)
+       !Implicit Integer (H-N)
        
-       Real (Kind=Kind(0.d0)), Dimension(:), allocatable :: XQMC, XTAU, Alpha_tot, om_bf, alp_bf, xom, A
+       Real (Kind=Kind(0.d0)), Dimension(:)  , allocatable :: XQMC, XTAU, Alpha_tot, om_bf, alp_bf, xom, A
        Real (Kind=Kind(0.d0)), Dimension(:,:), allocatable :: XCOV, Xn
-       Real (Kind=Kind(0.d0)) :: X_moments(2), Xerr_moments(2)
-       Real (Kind=Kind(0.d0)), External :: XKER, Back_trans_Aom
-       Character (Len=64) :: command, File1, File2
-       Complex (Kind=Kind(0.d0)) :: Z
+       Real (Kind=Kind(0.d0))                              :: X_moments(2), Xerr_moments(2)
+       Real (Kind=Kind(0.d0)), External                    :: XKER_ph, Back_trans_ph, XKER_pp, Back_trans_pp
+       Character (Len=64)                                  :: command, File1, File2
+       Complex (Kind=Kind(0.d0))                           :: Z
        
 
-
-       Open(unit=50,File='Info',Status="unknown")
+       Integer                :: Ngamma, Ndis,  NBins, NSweeps, Nwarm, N_alpha, L_cov
+       Real (Kind=Kind(0.d0)) :: OM_st, OM_en,  alpha_st, R
+       Logical                :: Checkpoint
+       Character (Len=2)      :: Channel
        
-       open(unit=30,file='paramSAC_ph',status='old',action='read', iostat=io_error) 
+       Integer                :: nt, nt1, io_error, n,nw, nwp, ntau, N_alpha_1, i
+       Real (Kind=Kind(0.d0)) :: dtau, pi, xmom1, x,x1,x2, tau, omp, om, Beta,err, delta, Dom
+
+       NAMELIST /VAR_Max_Stoch/ Ngamma, Ndis,  NBins, NSweeps, Nwarm, N_alpha, L_cov, &
+            &                   OM_st, OM_en,  alpha_st, R,  Checkpoint, Channel
+
+       open(unit=30,file='paramSAC',status='old',action='read', iostat=io_error) 
        if (io_error.eq.0) then
-          write(50,*) 'Reading in from paramSAC_ph'
-          read(30,*)  Ngamma, OM_st, OM_en, Ndis,  NBins, NSweeps, Nwarm
-          read(30,*)  N_alpha, alpha_st, R
-          read(30,*) L_cov
+          READ(30,NML=VAR_Max_Stoch)
        else
-          write(6,*) 'No file paramSAC_ph! ' 
+          write(6,*) 'No file paramSAC ' 
           stop
        endif
 
+       Open(unit=50,File='Info_MaxEnt',Status="unknown")
+       write(50,*) 'Channel      :: ', Channel
+       If (Channel == "ph" )  then
+          Write(50,*)  'Om_start is set to zero. PH channel corresponds to symmetric data '
+          Om_st = 0.d0
+       endif
+       Write(50,*) 'Covariance         :: ', L_cov
+       Write(50,*) 'Checkpoint         :: ', Checkpoint
+       Write(50,*) 'Om_st, Om_en       :: ', Om_st, Om_en
+       Write(50,*) 'Delta Om           :: ', (Om_en - Om_st)/real(Ndis,kind(0.d0))
+       Write(50,*) 'Bins, Sweeps, Warm :: ', NBins, NSweeps, Nwarm
+       If (N_alpha <= 10 ) then
+          Write(50,*) 'Not enough temperatures: N_alpha has to be bigger than 10'
+          Stop
+       Endif
+       Write(50,*) 'N_Alpha, Alpha_st,R:: ', N_alpha, alpha_st, R
+       
 
        open (unit=10,File="g_dat", status="unknown") 
-       read(10,*)  ntau
+       read(10,*)  ntau, n, Beta
        Allocate ( XCOV(NTAU,NTAU), XQMC(NTAU),XTAU(NTAU) )
        XCOV  = 0.d0
        Do nt = 1,NTAU
@@ -44,12 +67,21 @@
           enddo
        endif
        close(10)
-
        dtau = Xtau(2) - Xtau(1)
-       Beta = Xtau(Ntau)
 
        pi = acos(-1.d0)
-       xmom1 = pi * xqmc(ntau) 
+       Select Case (Channel)
+       Case ("PH")
+          xmom1 = pi * xqmc(1) 
+       Case ("PP")
+          xmom1 = 2.d0* pi * xqmc(1)
+       Case ("P")
+          xmom1 =  pi 
+       Case default 
+          Write(6,*) "Channel not yet implemented"
+          Stop
+       end Select
+
        Allocate (Alpha_tot(N_alpha) )
        do nt = 1,N_alpha
           alpha_tot(nt) = alpha_st*(R**(nt-1))
@@ -57,18 +89,39 @@
        write(50,*) 'First Moment, Beta  ', Xmom1, Beta
        close(50)
 
+       Select Case (Channel)
+       Case ("PH")
+          If (L_cov == 1 ) then
+             Call MaxEnt_stoch(XQMC, Xtau, Xcov, Xmom1, XKER_ph, Back_Trans_ph, Beta, &
+                  &            Alpha_tot, Ngamma, OM_ST, OM_EN, Ndis, Nsweeps, NBins, NWarm, L_Cov)
+          else
+             Call MaxEnt_stoch(XQMC, Xtau, Xcov, Xmom1, XKER_ph, Back_Trans_ph, Beta, &
+                  &            Alpha_tot, Ngamma, OM_ST, OM_EN, Ndis, Nsweeps, NBins, NWarm)
+          endif
+          ! Beware: Xqmc and cov are modified in the MaxEnt_stoch call.
+       Case ("PP")
+          If (L_cov == 1 ) then
+             Call MaxEnt_stoch(XQMC, Xtau, Xcov, Xmom1, XKER_pp, Back_Trans_pp, Beta, &
+                  &            Alpha_tot, Ngamma, OM_ST, OM_EN, Ndis, Nsweeps, NBins, NWarm, L_Cov)
+          else
+             Call MaxEnt_stoch(XQMC, Xtau, Xcov, Xmom1, XKER_pp, Back_Trans_pp, Beta, &
+                  &            Alpha_tot, Ngamma, OM_ST, OM_EN, Ndis, Nsweeps, NBins, NWarm)
+          endif
+          ! Beware: Xqmc and cov are modified in the MaxEnt_stoch call.
+       Case default 
+          Write(6,*) "Channel not yet implemented"
+          Stop
+       end Select
        
-       Call MaxEnt_stoch(XQMC, Xtau, Xcov, Xmom1, XKER, Back_Trans_Aom, Beta, &
-            &            Alpha_tot, Ngamma, OM_ST, OM_EN, Ndis, Nsweeps, NBins, NWarm)
-
-       Command = "rm dump*"
-       Call System (Command)
-       Command = "ls"
-       Call System (Command)
-
+       If ( .not.  Checkpoint ) then
+          Command = "rm dump*"
+          Call System (Command)
+          Command = "ls"
+          Call System (Command)
+       endif
        
        open (unit=10,File="g_dat", status="unknown") 
-       read(10,*)  ntau
+       read(10,*)  ntau, n, beta
        XCOV  = 0.d0
        Do nt = 1,NTAU
           read(10,*)  xtau(nt), xqmc(nt), err 
@@ -76,7 +129,7 @@
        Enddo
        close(10)
        pi = acos(-1.d0)
-       xmom1 = pi * xqmc(ntau) 
+       xmom1 = pi * xqmc(1) 
 
        Open (Unit = 10,File="Best_fit", Status ="unknown") 
        Allocate (om_bf(Ngamma), alp_bf(Ngamma) )
@@ -90,7 +143,7 @@
           X = 0.d0
           tau = xtau(nt)
           do i = 1,Ngamma
-             X = X + alp_bf(i)*XKER(tau,om_bf(i), beta)
+             X = X + alp_bf(i)*Xker_ph(tau,om_bf(i), beta)
           enddo
           Write(11,"(F14.7,2x,F14.7,2x,F14.7,2x,F14.7)")  xtau(nt), xqmc(nt),  sqrt(xcov(nt,nt)), xmom1*X
        enddo
@@ -99,7 +152,7 @@
        N_alpha_1 = N_alpha - 10  
        File1 ="Aom_ps"
        file2 =File_i(File1,N_alpha_1)
-       Write(50,*) file2
+       !Write(50,*) file2
        Open(Unit=66,File=file2,status="unknown")
        Allocate (xom(Ndis), A(Ndis))
        do nw = 1,Ndis
@@ -129,25 +182,75 @@
      
 
 
-     Real (Kind=Kind(0.d0)) function XKER(tau,om, beta)
+     Real (Kind=Kind(0.d0)) function XKER_ph(tau,om, beta)
 
        Implicit None
        real (Kind=Kind(0.d0)) :: tau, om, pi, beta
 
        pi = 3.1415927
 
-       XKER = (exp(-tau*om) + exp(-( beta - tau )*om ) )/( pi*(1.d0 + exp( - beta * om ) ) )
+       XKER_ph = (exp(-tau*om) + exp(-( beta - tau )*om ) )/( pi*(1.d0 + exp( - beta * om ) ) )
 
-     end function XKER
+     end function XKER_ph
 
-     Real (Kind=Kind(0.d0)) function Back_trans_Aom(Aom, om, beta)
+     Real (Kind=Kind(0.d0)) function XKER_pp(tau,om, beta)
+
+       Implicit None
+       real (Kind=Kind(0.d0)) :: tau, om, pi, beta
+
+       pi = 3.1415927
+
+       XKER_pp = exp(-tau*om) / ( pi*(1.d0 + exp( - beta * om ) ) )
+
+     end function XKER_pp
+
+     Real (Kind=Kind(0.d0)) function XKER_p(tau,om, beta)
+
+       Implicit None
+       real (Kind=Kind(0.d0)) :: tau, om, pi, beta
+
+       pi = 3.1415927
+
+       XKER_p  = exp(-tau*om) / ( pi*(1.d0 + exp( - beta * om ) ) )
+
+     end function XKER_p
+
+     
+     Real (Kind=Kind(0.d0)) function Back_trans_ph(Aom, om, beta)
 
        Implicit None
        real (Kind=Kind(0.d0)) ::  Aom, om, beta
 
-       Back_trans_Aom = Aom/(1.d0 + exp(-beta*om) )   ! This gives S(q,om) = chi(q,om)/(1 - e^(-beta om))
-       !Back_trans_Aom = Aom
+       Back_trans_ph = Aom/(1.d0 + exp(-beta*om) )   
+       ! This gives S(q,om) = chi(q,om)/(1 - e^(-beta om))
 
-     end function BACK_TRANS_AOM
+     end function BACK_TRANS_PH
+
+     Real (Kind=Kind(0.d0)) function Back_trans_pp(Aom, om, beta)
+            
+       Implicit None
+       real (Kind=Kind(0.d0)) ::  Aom, om, beta
+
+       real (Kind=Kind(0.d0)) :: Zero
+
+       Zero = 1.D-8
+
+       if ( abs(om) < zero ) then
+          Back_trans_pp = beta * Aom/2.d0
+       else
+          Back_trans_pp = Aom * (1.d0 + exp(-beta*om) ) / (om *( 1.d0 + exp(-beta*om) ) )
+       endif
+       ! This gives  = chi(q,om)/omega
+
+     end function BACK_TRANS_PP
+
+     Real (Kind=Kind(0.d0)) function Back_trans_p(Aom, om, beta)
+            
+       Implicit None
+       real (Kind=Kind(0.d0)) ::  Aom, om, beta
+
+       Back_trans_p =  Aom
+       
+     end function BACK_TRANS_P
